# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SEAPEDIA API",
    "version": "1.0.0",
    "description": "Multi-role e-commerce platform API. Supports Admin, Seller, Buyer, and Driver workflows. Protected endpoints require a Bearer JWT token in the Authorization header."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "description": "Creates a new account. Roles can be BUYER, SELLER, DRIVER. Returns a JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "username",
                  "email",
                  "password",
                  "roles"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "BUYER",
                        "SELLER",
                        "DRIVER"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful"
          },
          "400": {
            "description": "Email or username already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Returns JWT. The token includes user roles; call /auth/switch-role to set activeRole.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Account banned"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current user profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user with latest profile data"
          },
          "401": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/auth/profile": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Update profile",
        "description": "Updates name, email, username, picture URL, phone, gender, birthDate. Optional password change.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "profilePictureUrl": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "birthDate": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated, new token issued if email/username changed"
          }
        }
      }
    },
    "/auth/switch-role": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Switch active role",
        "description": "Issues a new JWT with the selected activeRole embedded. Required before accessing role-specific dashboards.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "BUYER",
                      "SELLER",
                      "DRIVER",
                      "ADMIN"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token with activeRole"
          },
          "403": {
            "description": "User does not own this role"
          }
        }
      }
    },
...
  }
}