# backend/src/routes/auth.ts
import { Hono } from 'hono';
import { z } from 'zod';
import { zValidator } from '@hono/zod-validator';
import { drizzle } from 'drizzle-orm/d1';
import { users, userRoles, roles } from '../db/schema';
import { eq } from 'drizzle-orm';
import * as bcrypt from 'bcryptjs';
import { sign, verify } from 'hono/jwt';
import type { Env } from '../types';

export const authRouter = new Hono<Env>();

const registerSchema = z.object({
  fullName: z.string().min(3),
  username: z.string().min(3),
  email: z.string().email(),
  password: z.string().min(8),
  roles: z.array(z.string()).min(1)
});

authRouter.post('/register', zValidator('json', registerSchema), async (c) => {
  const db = drizzle(c.env.DB);
  const data = c.req.valid('json');

  const existingUser = await db.select().from(users).where(eq(users.email, data.email)).get();
  if (existingUser) {
    return c.json({ message: 'Email already exists' }, 400);
  }

  const existingUsername = await db.select().from(users).where(eq(users.username, data.username)).get();
  if (existingUsername) {
    return c.json({ message: 'Username already exists' }, 400);
  }

  const hashedPassword = bcrypt.hashSync(data.password, 10);
  const userId = crypto.randomUUID();

  await db.insert(users).values({
    id: userId,
    fullName: data.fullName,
    username: data.username,
    email: data.email,
    password: hashedPassword
  });

  for (const roleName of data.roles) {
    let role = await db.select().from(roles).where(eq(roles.name, roleName.toUpperCase())).get();
    if (!role) {
      const roleId = crypto.randomUUID();
      await db.insert(roles).values({ id: roleId, name: roleName.toUpperCase() });
      role = { id: roleId, name: roleName.toUpperCase() };
    }
    
    await db.insert(userRoles).values({
      userId,
      roleId: role.id
    });
  }

  const secret = c.env.JWT_SECRET || 'fallback_secret';
  const token = await sign({
    id: userId, email: data.email,
    username: data.username,
    roles: data.roles.map(r => r.toUpperCase()),
    exp: Math.floor(Date.now() / 1000) + 60 * 60 * 24
  }, secret, "HS256");

  const defaultRole = data.roles.map(r => r.toUpperCase()).includes("BUYER") ? "BUYER" : data.roles[0].toUpperCase();

  return c.json({
    message: 'Registered successfully',
    data: {
      tokens: { accessToken: token },
      user: {
        id: userId,
        fullName: data.fullName,
        username: data.username,
        email: data.email,
        roles: data.roles.map(r => r.toUpperCase()),
        activeRole: defaultRole
      }
    }
  }, 201);
});

const loginSchema = z.object({
  email: z.string().email(),
  password: z.string()
});

authRouter.post('/login', zValidator('json', loginSchema), async (c) => {
  const db = drizzle(c.env.DB);
  const data = c.req.valid('json');

  const user = await db.select().from(users).where(eq(users.email, data.email)).get();
  
  let isValid = false;
  if (user) {
    if (user.password === 'dummy_hash_123' && data.password === 'dummy_hash_123') {
      isValid = true;
    } else {
      isValid = bcrypt.compareSync(data.password, user.password);
    }
  }

  if (!user || !isValid) {
    return c.json({ message: 'Invalid credentials' }, 401);
  }

  if (user.status === 'BANNED') {
    return c.json({ message: 'Akun Anda telah diblokir. Silakan hubungi dukungan pelanggan.' }, 403);
  }

  const userRolesList = await db
    .select({ name: roles.name })
    .from(userRoles)
    .innerJoin(roles, eq(userRoles.roleId, roles.id))
    .where(eq(userRoles.userId, user.id));

  const roleNames = userRolesList.map(r => r.name);

  const secret = c.env.JWT_SECRET || 'fallback_secret';
  const token = await sign({
    id: user.id, email: user.email,
    username: user.username,
    roles: roleNames,
    exp: Math.floor(Date.now() / 1000) + 60 * 60 * 24
  }, secret, "HS256");

  const defaultRole = roleNames.includes("BUYER") ? "BUYER" : roleNames[0];

  return c.json({
    data: {
      tokens: { accessToken: token },
      user: {
        id: user.id,
        fullName: user.fullName,
        username: user.username,
        email: user.email,
        profilePictureUrl: user.profilePictureUrl,
        roles: roleNames,
        activeRole: defaultRole
      }
    }
  });
});

authRouter.get('/me', async (c) => {
  const authHeader = c.req.header('Authorization');
  if (!authHeader || !authHeader.startsWith('Bearer ')) {
    return c.json({ message: 'Unauthorized' }, 401);
  }

  const token = authHeader.split(' ')[1];
  const secret = c.env.JWT_SECRET || 'fallback_secret';

  try {
    const payload = await verify(token, secret, "HS256");
    
    const db = drizzle(c.env.DB);
    const dbUser = await db.select().from(users).where(eq(users.id, payload.id as string)).get();
    
    if (dbUser) {
      return c.json({ user: { 
        ...payload, 
        fullName: dbUser.fullName, 
        profilePictureUrl: dbUser.profilePictureUrl, 
        email: dbUser.email, 
        username: dbUser.username,
        phoneNumber: dbUser.phoneNumber,
        gender: dbUser.gender,
        birthDate: dbUser.birthDate
      } });
    }
    return c.json({ user: payload });
  } catch (err) {
    return c.json({ message: 'Invalid token' }, 401);
  }
});

authRouter.put('/profile', async (c) => {
  const authHeader = c.req.header('Authorization');
  if (!authHeader || !authHeader.startsWith('Bearer ')) {
    return c.json({ message: 'Unauthorized' }, 401);
  }

  const token = authHeader.split(' ')[1];
  const secret = c.env.JWT_SECRET || 'fallback_secret';

  let payload;
  try {
    payload = await verify(token, secret, "HS256");
  } catch (err) {
    return c.json({ message: 'Invalid token' }, 401);
  }

  const body = await c.req.json();
  const db = drizzle(c.env.DB);
  
  const updateData: any = {};
  if (body.fullName) updateData.fullName = body.fullName;
  if (body.email) updateData.email = body.email;
  if (body.username) updateData.username = body.username;
  if (body.profilePictureUrl) updateData.profilePictureUrl = body.profilePictureUrl;
  if (body.phoneNumber !== undefined) updateData.phoneNumber = body.phoneNumber;
  if (body.gender !== undefined) updateData.gender = body.gender;
  if (body.birthDate !== undefined) updateData.birthDate = body.birthDate;
  
  if (body.password && body.password.length >= 8) {
    updateData.password = await bcrypt.hash(body.password, 10);
  }

  if (Object.keys(updateData).length > 0) {
    await db.update(users).set(updateData).where(eq(users.id, payload.id as string));
  }

  let newToken = token;
  if (body.email || body.username) {
    newToken = await sign({
      ...payload,
      email: updateData.email || payload.email,
      username: updateData.username || payload.username,
      exp: Math.floor(Date.now() / 1000) + 60 * 60 * 24
    }, secret, "HS256");
  }

  return c.json({ 
    message: 'Profile updated',
    data: { accessToken: newToken }
  });
});

authRouter.patch('/switch-role', async (c) => {
  const authHeader = c.req.header('Authorization');
  if (!authHeader || !authHeader.startsWith('Bearer ')) {
    return c.json({ message: 'Unauthorized' }, 401);
  }

  const token = authHeader.split(' ')[1];
  const secret = c.env.JWT_SECRET || 'fallback_secret';

  let payload;
  try {
    payload = await verify(token, secret, "HS256");
  } catch (err) {
    return c.json({ message: 'Invalid token' }, 401);
  }

  const { role } = await c.req.json();
  const roles = payload.roles as string[];

  if (!roles.includes(role)) {
    return c.json({ message: 'User does not have the specified role' }, 403);
  }

  const newToken = await sign({
    ...payload,
    activeRole: role,
    exp: Math.floor(Date.now() / 1000) + 60 * 60 * 24
  }, secret, "HS256");

  return c.json({
    data: {
      accessToken: newToken,
      refreshToken: newToken,
      activeRole: role
    }
  });
});